1.8.0 Add Simplex Noise Generator. Needs more work. Add method "setBiomeProviderScript" to allow for a script based biome provider to be set for the dimension. This is directly lifted from DimensionalControl. Also added the gen layers for it. Added methods "addBiomes", "addOceanBiomes", "addHills" and "setRiver" as part of simplified multi-biome dimension generation. Add examples "Desert Overworld", "Thick Rivers" and "Repeating Heat Map"

+++ src/main/java/com/bloodnbonesgaming/topography/world/generator/CellNoiseGenerator.java
@@ -69,7 +69,7 @@ public class CellNoiseGenerator implements IGenerator
     }
     
     @ScriptMethodDocumentation(args = "float", usage = "cutoff", notes = "Sets the noise cutoff for blocks to be placed. Higher values = more blocks. Default is -0.15.")
-	public void setCutoff(final float cutoff)
+	public void setCutoff(final double cutoff)
     {
     	this.cutoff = cutoff;
     }
